noteTaking/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
crispy-bootstrap5==2023.10
Django==4.2.7
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-quill-editor==0.1.40
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==21.0.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
